% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuspectral_wrappers.R
\name{freq_axis}
\alias{freq_axis}
\title{Freq axis}
\usage{
freq_axis(t)
}
\arguments{
\item{t}{vector of time points}
}
\value{
freqs vector of analysis frequencies
}
\description{
Define frequency axis corresponding to time axis t
}
\seealso{
Other spectra: 
\code{\link{nupsd}()},
\code{\link{nuwavelet_psd}()},
\code{\link{nuwavelet}()}
}
\concept{spectra}
