% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuspectral_wrappers.R
\name{nupsd}
\alias{nupsd}
\title{nupsd}
\usage{
nupsd(time, vals, freqs = NA)
}
\arguments{
\item{time}{vector of time points}

\item{vals}{vector of values vals = y(time)}

\item{freqs}{vector of analysis frequencies}
}
\value{
psd
}
\description{
Lomb-Scargle periodogram derived from the `nuspectral` nurealcoeff routine
}
\references{
Matthias et al, (2004), Algorithms for Spectral Analysis of Irregularly Sampled Time Series, J. Stat. Soft.
}
\seealso{
Other spectra: \code{\link{freq_axis}},
  \code{\link{nuwavelet_psd}}, \code{\link{nuwavelet}}
}
