\name{nuspectral-package}
\alias{nuspectral-package}
\alias{nuspectral}
\docType{package}
\title{
\packageTitle{nuspectral}
}
\description{
\packageDescription{nuspectral}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{nuspectral}
\packageIndices{nuspectral}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{nuspectral}

Maintainer: \packageMaintainer{nuspectral}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%~~ simple examples of the most important functions ~~
}
