% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuspectral_wrappers.R
\name{nuwavelet}
\alias{nuwavelet}
\title{Scalogram}
\usage{
nuwavelet(time, vals, freqs, taus, wgt = cubicwgt, wgtrad = 1,
  sigma = 0.05)
}
\arguments{
\item{time}{vector of time points}

\item{vals}{vector of values vals = y(time)}

\item{freqs}{vector of analysis frequencies}

\item{taus}{vector of analysis time shifts}

\item{wgt}{weight function for the wavelet}

\item{wgtrad}{radius (scale) of the weight function}

\item{sigma}{scaling parameter of the wavelet}
}
\value{
a list of ensemble spectra results
\itemize{
\item sclgrm: 2D array (nfreqs x ntaus) of wavelet coefficients
\item Neffs: Number of effective degrees of freedom  (same dimensions)
}
}
\description{
Wrapper for the R function `nuwaveletcoeff`
}
\references{
Matthias et al, (2004), Algorithms for Spectral Analysis of Irregularly Sampled Time Series, J. Stat. Soft.

Foster, G. (1996), Wavelets for period analysis of unevenly sampled time series, Astron. Jour., 112, 1709, doi:10.1086/118137.
}
\seealso{
Other spectra: \code{\link{freq_axis}}
}
\concept{spectra}
