% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuspectral_wrappers.R
\name{nuwavelet_psd}
\alias{nuwavelet_psd}
\title{nuwavelet_psd}
\usage{
nuwavelet_psd(time, vals, freqs = NA, taus = NA, wgtrad = 1,
  sigma = 0.05)
}
\arguments{
\item{time}{vector of time points}

\item{vals}{vector of values vals = y(time)}

\item{freqs}{vector of analysis frequencies}

\item{taus}{vector of analysis time shifts}

\item{wgtrad}{radius (scale) of the weight function}

\item{sigma}{scaling parameter of the wavelet}

\item{wgt}{weight function for the wavelet}
}
\value{
psd
}
\description{
Power spectral density derived from a `nuspectral` scalogram
}
\references{
Matthias et al, (2004), Algorithms for Spectral Analysis of Irregularly Sampled Time Series, J. Stat. Soft.

Foster, G. (1996), Wavelets for period analysis of unevenly sampled time series, Astron. Jour., 112, 1709, doi:10.1086/118137.

Kirchner, J. W. & Neal, C. (2013), Universal fractal scaling in stream chemistry and its implications for solute transport and water quality trend detection. PNAS 110, 12213–12218.
}
\seealso{
Other spectra: \code{\link{freq_axis}},
  \code{\link{nupsd}}, \code{\link{nuwavelet}}
}
